refactor(proxy): simplify and speed up
