+++ apps/proxy/package.json
@@ -1,12 +1,11 @@
 {
-  "name": "proxy",
+  "name": "fxc-proxy",
   "version": "1.0.0",
   "description": "Forward fetcher requests",
   "type": "module",
   "main": "./index.js",
   "dependencies": {
-    "@flyxc/common": "workspace:*",
-    "express": "catalog:default"
+    "proxy": "catalog:default"
   },
   "devDependencies": {
     "@dotenvx/dotenvx": "catalog:default"
