+++ libs/common/src/protos/index.ts
@@ -1,4 +1,3 @@
 export * from './fetcher-state';
 export * from './live-track';
-export * from './proxy';
 export * from './track';
