+++ pnpm-workspace.yaml
@@ -39,9 +39,11 @@ catalog:
   jsonc-eslint-parser: 3.3.0
   lit: ^3.3.3
   mapbox-vector-tile: ^0.3.0
+  proxy: ^4.1.0
   simplify-path: ^1.1.0
   tiny-lru: ^13.0.0
   tslib: ^2.8.1
+  undici: ~7.29.0
   vite: ^8.2.2
   zod: ^4.5.4
   core-js: ^3.50.0
